fix: use original gradle/maven repo so we can get rid of the new one
Had created a second maven repo on artifactory because we couldn't
figure out how to set up creds to publish to the gradle one.  That
is a bad user experience because it means customers have to define
two repositories in their pom (one for this project, one for momento
SDK).

This commit fixes the pom to publish to the original gradle repo,
and also adds a README with example settings file.